refactor: see full commit message
- flattened packages to make the public API smoother
- renamed some packages to remove pluralities for consistency
- improved documentation for public API, especially for the
  debug combinators
- cleared context with each parser run so debugged parsers can
  be re-used freely as long as the trees are saved or discarded
  appropriately between runs
