+++ parsley-debug/shared/src/main/scala/parsley/debugger/internal/DebugTreeBuilder.scala
@@ -1,0 +1,29 @@
+package parsley.debugger.internal
+
+import parsley.internal.deepembedding.frontend.LazyParsley
+
+// Helper class for reconstructing a debug tree.
+// Not meant to be public.
+private [parsley] case class DebugTreeBuilder(
+  node: TransientDebugTree,
+  bChildren: Map[LazyParsley[Any], DebugTreeBuilder] = Map.empty,
+) {
+  def addNode(path: List[LazyParsley[Any]], node: TransientDebugTree): DebugTreeBuilder =
+    path match {
+      case Nil      => DebugTreeBuilder(node, Map.empty)
+      case p :: ps  =>
+        // Pre: The path to this node must fully exist.
+        // Tip: Add the shortest paths first!
+        val child = this.bChildren.getOrElse(p, DebugTreeBuilder(node))
+        DebugTreeBuilder(this.node, this.bChildren.+((p, child.addNode(ps, node))))
+    }
+
+  def reconstruct: TransientDebugTree = {
+    node.children
+      .addAll(
+        bChildren.map { case (lp, cs) => (Rename(lp), cs.reconstruct) }
+      )
+
+    node
+  }
+}
