+++ parsley-debug/shared/src/main/scala/parsley/internal/deepembedding/frontend/debugger/debugger.scala
@@ -3,8 +3,7 @@
  */
 package parsley.internal.deepembedding.frontend
 
-import parsley.debugger.objects.DebugContext
-
+import parsley.debugger.internal.DebugContext
 import parsley.internal.deepembedding.backend.StrictParsley
 import parsley.internal.deepembedding.frontend
